Move current_runtime to runtime init
