+++ src/dataclay/__init__.py
@@ -2,7 +2,6 @@
 Public dataclay functions exported to use (friendly) "from dataclay import ..."
 """
 from dataclay.runtime.Initializer import *
-from dataclay.runtime.Runtime import get_runtime
 from dataclay.DataClayObject import DataClayObject
 from dataclay.DataClayObjMethods import dclayEmptyMethod, dclayMethod
