+++ src/dataclay/communication/grpc/paraver/ParaverClientInterceptor.py
@@ -8,7 +8,7 @@ from logging import getLogger
 from grpc import UnaryUnaryClientInterceptor
 
 from dataclay import PrvManager
-from dataclay.runtime.settings import settings
+from dataclay.runtime import settings
 from dataclay.communication.grpc.paraver import HEADER_CLIENTPORT
 from dataclay.paraver.prv_traces import TraceType
