+++ src/dataclay/executionenv/ExecutionEnvironment.py
@@ -5,9 +5,8 @@ import pickle
 import traceback
 import uuid
 
+from dataclay.runtime import set_runtime, settings
 from dataclay.runtime.execution_environment_runtime import ExecutionEnvironmentRuntime
-from dataclay.runtime.Runtime import get_runtime, set_runtime
-from dataclay.runtime.settings import settings
 from dataclay.DataClayObject import DataClayObject
 from dataclay.DataClayObjProperties import (
     DCLAY_GETTER_PREFIX,
