+++ src/dataclay/loader/ExecutionObjectLoader.py
@@ -11,7 +11,7 @@ import time
 import traceback
 
 from dataclay.runtime.ExecutionGateway import ExecutionGateway
-from dataclay.runtime.Runtime import get_runtime
+from dataclay.runtime import get_runtime
 from dataclay.communication.grpc.Utils import get_metadata
 from dataclay.loader.DataClayObjectLoader import DataClayObjectLoader
 from dataclay.serialization.lib.DeserializationLibUtils import DeserializationLibUtilsSingleton
