+++ src/dataclay/util/StubUtils.py
@@ -5,8 +5,7 @@ import os
 
 from dataclay_common.protos.common_messages_pb2 import LANG_PYTHON
 
-from dataclay.runtime.Runtime import get_runtime
-from dataclay.runtime.settings import settings
+from dataclay.runtime import get_runtime, settings
 from dataclay.util.FileUtils import deploy_class
 from dataclay.util.YamlParser import Loader, dataclay_yaml_load
